Changed cpu calculation

+++ include/processor.h
@@ -8,16 +8,16 @@ class Processor {
   // TODO: Declare any necessary private members
  private:
 
-  int prevuser = 0;
-  int prevnice = 0;
-  int prevsystem = 0;
-  int previdle = 0;
-  int previowait = 0;
-  int previrq = 0;
-  int prevsoftirq = 0;
-  int prevsteal = 0;
-  int prevguest = 0;
-  int prevguest_nice = 0;
+  float prevuser = 0;
+  float prevnice = 0;
+  float prevsystem = 0;
+  float previdle = 0;
+  float previowait = 0;
+  float previrq = 0;
+  float prevsoftirq = 0;
+  float prevsteal = 0;
+  float prevguest = 0;
+  float prevguest_nice = 0;
 };
 
 #endif
